% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesQVGEL.R
\name{BayesQVGEL}
\alias{BayesQVGEL}
\title{fit a Bayesian quantile variable selection for g - e in longitudinal studies}
\usage{
BayesQVGEL(
  y,
  e,
  C,
  g,
  w,
  k,
  iterations = 10000,
  burn.in = NULL,
  slope = TRUE,
  robust = TRUE,
  quant = 0.5,
  sparse = TRUE,
  structure = c("group", "individual")
)
}
\arguments{
\item{y}{the matrix of response variable. The current version of BayesQVGEL only supports continuous response.}

\item{e}{the matrix of a group of dummy environmental factors variables.}

\item{C}{the matrix of the intercept and time effects(time effects are optional).}

\item{g}{the matrix of predictors (genetic factors) without intercept. Each row should be an observation vector.}

\item{w}{the matrix of interactions between genetic factors and environmental factors.}

\item{k}{the total number of time points.}

\item{iterations}{the number of MCMC iterations.}

\item{burn.in}{the number of iterations for burn-in.}

\item{slope}{logical flag. If TRUE, random intercept and slope model will be used.}

\item{robust}{logical flag. If TRUE, robust methods will be used.}

\item{quant}{specify different quantiles when applying robust methods.}

\item{sparse}{logical flag. If TRUE, spike-and-slab priors will be used to shrink coefficients of irrelevant covariates to zero exactly.}

\item{structure}{structure for interaction effects, two choices are available. "group" for selection on group-level only. "individual" for selection on individual-level only.}
}
\value{
an object of class `BayesQVGEL' is returned, which is a list with component:
\item{posterior}{the posteriors of coefficients.}
\item{coefficient}{the estimated coefficients.}
\item{burn.in}{the total number of burn-ins.}
\item{iterations}{the total number of iterations.}
}
\description{
fit a Bayesian quantile variable selection for g - e in longitudinal studies
}
\details{
Consider the data model described in "\code{\link{data}}":
\deqn{Y_{ij} = X_{ij}^{T}\gamma_{0}+E_{ij}^{T}\gamma_{1}+\sum_{l=1}^{p}G_{ijl}\gamma_{2l}+\sum_{l=1}^{p}W_{ijl}^{T}\gamma_{3l}+Z_{ij}^{T}\alpha_{i}+\epsilon_{ij}.}
where \eqn{\gamma_{2l}} is the main effect of the \eqn{l}th genetic variant. The interaction effects is corresponding to the coefficient vector \eqn{\gamma_{3l}=(\gamma_{3l1}, \gamma_{3l2},\ldots,\gamma_{3lm})^\top}.

When {structure="group"}, group-level selection will be conducted on \eqn{||\gamma_{3l}||_{2}}. If {structure="individual"}, individual-level selection will be conducted on each \eqn{\gamma_{3lq}}, (\eqn{q=1,\ldots,m}).

When {slope=TRUE} (default), random intercept and slope model will be used as the mixed effects model.

When {sparse=TRUE} (default), spike--and--slab priors are imposed on individual and/or group levels to identify important main and interaction effects. Otherwise, Laplacian shrinkage will be used.

When {robust=TRUE} (default), the distribution of \eqn{\epsilon_{ij}} is defined as a Laplace distribution with density.

\eqn{
f(\epsilon_{ij}|\theta,\tau) = \theta(1-\theta)\exp\left\{-\tau\rho_{\theta}(\epsilon_{ij})\right\}
}, (\eqn{i=1,\dots,n,j=1,\dots,k }), which leads to a Bayesian formulation of quantile regression. If {robust=FALSE}, \eqn{\epsilon_{ij}} follows a normal distribution.
\cr

Please check the references for more details about the prior distributions.
}
\examples{
data(data)

## default method
fit=BayesQVGEL(y,e,C,g,w,k,structure=c("group"))
fit$coefficient

## Compute TP and FP
b = selection(fit,sparse=TRUE)
index = which(coeff!=0)
pos = which(b != 0)
tp = length(intersect(index, pos))
fp = length(pos) - tp
list(tp=tp, fp=fp)
\donttest{
## alternative: robust individual selection
fit=BayesQVGEL(y,e,C,g,w,k,structure=c("individual"))
fit$coefficient

## alternative: non-robust group selection
fit=BayesQVGEL(y,e,C,g,w,k,robust=FALSE, structure=c("group"))
fit$coefficient

## alternative: robust group selection under random intercept model
fit=BayesQVGEL(y,e,C,g,w,k,slope=FALSE, structure=c("group"))
fit$coefficient

}

}
\seealso{
\code{\link{data}}
}
\keyword{models}
