% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE.R
\name{GE}
\alias{GE}
\title{Construct Gene-Environment (G×E) Interaction Matrix}
\usage{
GE(g,e)
}
\arguments{
\item{g}{the long-format matrix of genetic predictors.}

\item{e}{the long-format design matrix for environment/treatment effects.}
}
\value{
the G×E interaction terms.
}
\description{
Construct Gene-Environment (G×E) Interaction Matrix
}
