% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\title{This function changes the format of the longitudinal data from wide format to long format}
\usage{
reformat(k,data,type=c("r","d"))
}
\arguments{
\item{k}{the number of repeated measurement.}

\item{data}{either the response matrix or the predictor matrix.}

\item{type}{"r" for response, "d" for design.}
}
\value{
the reformatted response vector or predictor matrix.
}
\description{
This function changes the format of the longitudinal data from wide format to long format
}
