\name{rolong-package}
\alias{rolong-package}
\alias{rolong}
\docType{package}
\title{
\packageTitle{rolong}
}
\description{
\packageDescription{rolong}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rolong}
\packageIndices{rolong}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{rolong}

Maintainer: \packageMaintainer{rolong}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
