% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{reformat}
\alias{reformat}
\title{This function changes the format of the longitudinal data from wide format to long format}
\usage{
reformat(k,y,x)
}
\arguments{
\item{k}{the number of repeated measurement.}

\item{y}{the matrix of longitudinal response.}

\item{x}{the matrix of predictors.}
}
\value{
a list containing the reformatted response vector and predictor matrix.
}
\description{
This function changes the format of the longitudinal data from wide format to long format
}
