% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolong-package.R
\docType{package}
\name{rolong-package}
\alias{rolong-package}
\title{Bayesian quantile variable selection with mixed effects model in longitudinal study}
\description{
In longitudinal studies, the same subjects are measured repeatedly over time, leading to correlations among the repeated measurements. Properly accounting for the intra-cluster correlations in the presence of data heterogeneity and long tailed distributions of the disease phenotype is challenging, especially in the context of high dimensional regressions. Here, we aim at developing novel Bayesian regularized quantile mixed effect models to tackle these challenges.In this package, we have implemented the Gibbs samplers of Bayesian quantile variable selection method in the mixed effects model with the spike-and-slab priors. The Markov Chain Monte Carlo (MCMC) algorithms of the proposed and alternative models can be efficiently performed by using the package.
}
\details{
The user friendly, integrated interface \strong{rolong()} allows users to flexibly choose the fitting methods by specifying the following parameter:
\tabular{rl}{
sparse: \tab whether to use the spike-and-slab priors to impose sparsity.
}

The function rolong() returns a rolong object that contains the posterior estimates of each coefficients.
}
\references{
Zhou, F., Ren, J., Ma, S. and Wu, C. (2022). The Bayesian regularized quantile varying coefficient model. (submitted)

Ren, J., Zhou, F., Li, X., Ma, S., Jiang, Y. and Wu, C. (2022). Robust Bayesian variable selection for gene-environment interactions.
{\emph{Biometrics}, (in press)} \doi{10.1111/biom.13670}

Wu, C., and Ma, S. (2015). A selective review of robust variable selection with applications in bioinformatics.
{\emph{Briefings in Bioinformatics}, 16(5), 873–883} \doi{10.1093/bib/bbu046}

Zhou, F., Ren, J., Lu, X., Ma, S. and Wu, C. (2021). Gene–Environment Interaction: a Variable Selection Perspective.
{\emph{Epistasis. Methods in Molecular Biology.} 2212:191–223} \url{https://link.springer.com/protocol/10.1007/978-1-0716-0947-7_13}

Ren, J., Zhou, F., Li, X., Chen, Q., Zhang, H., Ma, S., Jiang, Y. and Wu, C. (2020) Semi-parametric Bayesian variable selection for gene-environment interactions.
{\emph{Statistics in Medicine}, 39: 617– 638} \doi{10.1002/sim.8434}

Ren, J., Zhou, F., Li, X., Wu, C. and Jiang, Y. (2019) spinBayes: Semi-Parametric Gene-Environment Interaction via Bayesian Variable Selection.
R package version 0.1.0. \url{https://CRAN.R-project.org/package=spinBayes}

Wu, C., Jiang, Y., Ren, J., Cui, Y. and Ma, S. (2018). Dissecting gene-environment interactions: A penalized robust approach accounting for hierarchical structures.
{\emph{Statistics in Medicine}, 37:437–456} \doi{10.1002/sim.7518}

Wu, C., Shi, X., Cui, Y. and Ma, S. (2015). A penalized robust semiparametric approach for gene-environment interactions.
{\emph{Statistics in Medicine}, 34 (30): 4016–4030} \doi{10.1002/sim.6609}

Wu, C., Cui, Y., and Ma, S. (2014). Integrative analysis of gene–environment interactions under a multi–response partially linear varying coefficient model.
{\emph{Statistics in Medicine}, 33(28), 4988–4998} \doi{10.1002/sim.6287}

Wu, C., Zhong, P.S. and Cui, Y. (2018). Additive varying–coefficient model for nonlinear gene–environment interactions.
{\emph{Statistical Applications in Genetics and Molecular Biology}, 17(2)} \doi{10.1515/sagmb-2017-0008}

Wu, C., Zhong, P.S. and Cui, Y. (2013). High dimensional variable selection for gene-environment interactions.
{\emph{Technical Report. Michigan State University.}}
}
\seealso{
\code{\link{rolong}}
}
\keyword{overview}
